API to read entire trace point at once.
This can be done with minimal allocation by using buffers in the state
struct. Since the buffers are reused, the caller must process or copy
the entire point before the next call to read_trace_point.

This is significantly faster (about a factor of four) to use from Common
Lisp due to the per-call overhead of CFFI.