Add helper functions for writing variables.
WRITE_TRACE_VARIABLE is a fairly large macro which can be invoked
thousands of times in an instrumented program, leading to long compile
times. Moving most of it into a function cuts down on the bloat and
speeds things up.

We still need the macro to create a temporary variable (for writing
register variables).

write_trace_blob doesn't need to deal with register variables so it
can be implemented entirely as a function.