Bumping the size of various fields.
This will allow support for larger projects, such as
wireshark.

We now allow 2^31 variable names and types.  Additionally,
a variable size can be as much as 2^31 bits.

The statement id is now 2^64 bits (23 bits for the file and 40
for the AST number).

Most importantly, the first field in the trace, the number of
characters in all instrumented variables, is now 2^64 bits.
This value would previously overflow on large programs, such
as wireshark.