Use arbitrarily large integers in preconditions.
When evaluating preconditions, use arbitrarily large
integers.  This allows for easier expression of
integer overflow/underflows.  Additionally, the previous
implementation, using the `int_value' struct was incorrect
in an irreparable manner using non-arbitrary integers.
For instance, when evaluating the following expression,
the `int_value' returned would be 0, not -1, due to C/C++
rules for type casting: -2147483647u / -1.
This led to errors in injection of underflow scions.

The new implementation uses the `cint' classes from boost,
necessitating the inclusion of a large number of boost-related
modules as dependencies.  Unfortunately, I am not aware of
similar library relating to arbitrarily large integers
which is (1) professionally written, (2) has a painless to use
license, and (3) is header-only relating only.  Thus, the
sledgehammer approach was taken.

These boost dependencies are from version 1.66.