Build Boost as part of trace database.
If a system install of Boost is found, we will use this.
Otherwise, we will download Boost and build the serialization
library in third-party.

The client may also specify a BOOST_INCLUDE_DIR and BOOST_LIB_DIR
in a local.mk file if Boost is installed in a non-standard location.