Improve space/time performance via flyweights
This commit improves upon the existing trace database by
utilizing the flyweight design pattern to share memory
usage.

While this commit touches many files, this should not be
mistaken for a large change in behavior.  In fact, the
behavior of the trace database after these commits
should remain unchanged.  Great effort was given to ensure
behavior was preserved; while it is believed that the existing
behavior is correct, any pre-existing flaws may have been
propogated forward.  The underlying trace format remains
unchanged.

This commit restuctures the trace database into a header-only C++
library.  This header-only library is defined in the hpp
files found at the top level of the project.  The existing sample.cpp
file, useful for debugging, remains at the top level and leverages
this C++ library.  The hope is that this header-only library
may be easily utilized by clients looking to bring the
trace database into other projects.

Beyond this, the "lisp/" directory contains a C wrapper for the
trace database.  The wrapper defines the functions and C data
structures declared in binary-trace-db.lisp.  The binary trace db
is the lisp object utilized to interact with the trace database.

As discussed above, the trace database now utilizes the flyweight
design pattern (via the the boost flyweight library) to reduce
memory usage and improve query performance.  Query results
are cached, so duplicate trace points are now queried only once.

Beyond this, effort was made to reduce the number of
vector copies required for queries.  Additionally, when no
filtering is done on the common lisp side (e.g. when predicates
are used), result deduplication is done on the C side, improving
performance.  Support for floating point numbers and additional
predicates, such as <=, =, and >=, were also added.

With these changes, the following performance numbers can
be reported.  The statistics were collected using a single
grep trace containing approximately 86000 points.

The first table shows the time to parse the trace, the
second, the amount of memory required for the trace, and the last
four the amount of time required to query the trace for specific
Bug-Injector scions.

| Variant              | Parse Time (s) |
|----------------------|----------------|
| original             |           0.66 |
| c++ w/o flyweights   |           1.07 |
| c++ w/ flyweights    |           1.19 |

| Variant      |  Memory Usage (MB) |
|--------------|--------------------|
| original     |              50.52 |
| flyweights   |              26.88 |

arithmetic-overflow-multiplication-int
| Variant      | Query Time (s) |
|--------------|----------------|
| original     |           6.15 |
| flyweights   |           1.13 |

arithmetic-overflow-addition-int
| Variant      | Query Time (s) |
|--------------|----------------|
| original     |          40.81 |
| flyweights   |          14.48 |

division-by-zero-int
| Variant      | Query Time (s) |
|--------------|----------------|
| original     |         272.67 |
| flyweights   |          41.64 |

buf-overflow-memcpy
| Variant      | Query Time (s) |
|--------------|----------------|
| original     |         116.93 |
| flyweights   |           5.63 |