Reduce trace database memory requirements
This changeset reduces the trace database memory requirements
for query result caching by at least an order of magnitude.  There
are several items which contributed to this memory usage reduction.

1. We always call `shrink_to_fit` before caching vectors.
2. We introduce a functor, `EvaluatePredicate`, which evaluates
whether a predicate is satisfied for a vector of flyweight
trace variables.  We cache the results of the functor.
Since variables often have the same value (e.g. globals
initialized to zero) across multiple points, this yields
a non-trivial reduction in memory usage.
3. Instead of having a unique cache for predicate results on each
trace point, we now use a single cache.

These changes do not have any noticeable impact on the timing results
reported in a3e7ad3b.

-----------------------------------------------------------------------

Statistics for grep with five traces searching for division-by-zero
scions:

| Variant  | Memory Usage (Percentage of Machine Total) |
|----------|--------------------------------------------|
| original | 14.47%                                     |
| branch   |  0.91%                                     |

Given that around 0.6% of memory is utilized after trace collection,
this is a *substantial* reduction in memory required for query
caching.